/produtos path post e delete respondem json

+++ backend/src/routes/produtos.js
@@ -1,7 +1,8 @@
 const express = require("express")
 const router = express.Router()
 const model_produto = require("../models/produto");
-const {v4:uuidv4}= require("uuid")
+const {v4:uuidv4}= require("uuid");
+const cliente = require("../models/cliente");
 
 router.get('/',async(req,res)=>{
 
@@ -27,9 +28,15 @@ router.post('/', async(req,res)=>{
     try{
        const {sabor,preco,tipo,lucro} =  req.body
        const id = uuidv4()
-       produto = {sabor,preco,tipo,id,lucro}            
+       if(lucro == null){
+        produto = {sabor,preco,tipo,id} 
+       }
+       else{
+        produto = {sabor,preco,tipo,id,lucro} 
+       }
+        produto = await model_produto.findByPk(id)
         await model_produto.create(produto)
-        res.send(200)
+        res.status(200).send(produto)
     }
     catch(err){
         console.log(err)
@@ -40,12 +47,17 @@ router.post('/', async(req,res)=>{
 
 router.patch('/', async(req,res)=>{
   try{
-     const {sabor,preco,tipo,id,lucro}  =  req.body  
-     if(!model_produto.findByPk(id)){
+     const {sabor,preco,tipo,id,lucro}  =  req.body 
+     let produto = model_produto.findByPk(id)
+     if(!id){
+        throw new Error()
+     }
+     if(!produto){
       throw new Error()
      }
-     model_produto.update({sabor:sabor,preco:preco,tipo:tipo,lucro:lucro},{where:{id:id}})
-     res.send(200)
+     await model_produto.update({sabor:sabor,preco:preco,tipo:tipo,lucro:lucro},{where:{id:id}})
+     produto = await model_produto.findByPk(id)
+     res.status(200).send(produto)
   }
   catch(err){
       console.log(err)
@@ -57,8 +69,15 @@ router.patch('/', async(req,res)=>{
 router.delete('/', async(req,res)=>{
   try{
      const {id} =  req.body  
+     let produto = await model_produto.findByPk(id)
+     if(!id){
+        throw new Error()
+     }
+     else if(!cliente){
+        throw new Error()
+     }
      await model_produto.destroy({where:{id:id}})
-    res.send(200)
+    res.status(200).send(produto)
   }
   catch(err){
       console.log(err)
